refactor: handler into a single file
